move prob distribution
