+++ probnode/__init__.py
@@ -3,4 +3,4 @@ from .probability.event import *
 from .probability.sample_space import SampleSpace
 from .probability.random_variable import RandomVariable
 from .probability.probability_measure import ProbabilityMeasureOfEvent, P__
-from .datatype import ProbabilityDistribution
+from .probability.probability_distribution import ProbabilityDistribution
