+++ probnode/datatype/__init__.py
@@ -1,2 +1,1 @@
 from .probability_value import *
-from .probability_distribution import ProbabilityDistribution
